+++ src/main/java/net/geforcemods/securitycraft/blocks/mines/MineBlock.java
@@ -74,7 +74,7 @@ public class MineBlock extends ExplosiveBlock {
 			return;
 		else if (!getShape(state, level, pos, CollisionContext.of(entity)).bounds().move(pos).inflate(0.01D).intersects(entity.getBoundingBox()))
 			return;
-		else if (!EntityUtils.doesEntityOwn(entity, level, pos))
+		else if (!EntityUtils.doesEntityOwn(entity, level, pos) && !(entity instanceof Player player && player.isCreative()))
 			explode(level, pos);
 	}